Created a participant-suitability.model and moved the common properties here.
Modified the individual-suitability.model to extend the participant-suitability.model.
