+++ ServiceWebsite/ServiceWebsite/ClientApp/src/app/modules/individual-journey/individual-suitability.model.ts
@@ -1,41 +1,10 @@
-export class SuitabilityAnswer {
-    answer: boolean;
-    notes: string;
-}
-
-export enum HasAccessToCamera {
-    Yes,
-    No,
-    NotSure
-}
+import { ParticipantSuitabilityModel, SuitabilityAnswer } from '../base-journey/participant-suitability.model';
 
 /**
  * Exposes the basic properties of the suitability model.
  */
-export abstract class IndividualSuitabilityModel {
-    camera: HasAccessToCamera;
-
-    computer: boolean;
+export abstract class IndividualSuitabilityModel extends ParticipantSuitabilityModel {
     internet: boolean;
     interpreter: boolean;
-    room: boolean;
-
-    aboutYou: SuitabilityAnswer;
     consent: SuitabilityAnswer;
-
-    hearing: Hearing;
-
-    isUpcoming(): boolean {
-        const now = new Date();
-        return this.hearing.scheduleDateTime >= now;
-    }
-}
-
-export class Hearing {
-    constructor(id?: string, scheduledDateTime?: Date) {
-        this.id = id;
-        this.scheduleDateTime = scheduledDateTime;
-    }
-    id: string;
-    scheduleDateTime: Date;
 }
