+++ ServiceWebsite/ServiceWebsite/ClientApp/src/app/modules/representative-journey/representative-suitability.model.ts
@@ -1,0 +1,10 @@
+import { ParticipantSuitabilityModel, SuitabilityAnswer } from '../base-journey/participant-suitability.model';
+
+/**
+ * Exposes the basic properties of the suitability model.
+ */
+export abstract class RepresentativeSuitabilityModel extends ParticipantSuitabilityModel {
+    aboutClient: SuitabilityAnswer;
+    clientAttenance: boolean;
+    hearingSuitability: SuitabilityAnswer;
+}
